chore(use-acl-columns): add 'renderRoleActions' [YTFRONT-5158]
